refactored general purpose types into a separate module
